requirements.txt
streamlit
pandas
numpy
plotly
pillow
pytesseract
pdfplumber
scikit-learn
openpyxl
